3619: Rename RemoveEntityAttributesQuickFix files
