+++ common/src/Model/LongPropertyValueIssueGenerator.cpp
@@ -20,13 +20,13 @@
 #include "LongPropertyValueIssueGenerator.h"
 
 #include "Model/BrushNode.h"
-#include "Model/RemoveEntityAttributesQuickFix.h"
 #include "Model/Entity.h"
 #include "Model/EntityNode.h"
 #include "Model/Issue.h"
 #include "Model/IssueQuickFix.h"
 #include "Model/MapFacade.h"
 #include "Model/PushSelection.h"
+#include "Model/RemoveEntityPropertiesQuickFix.h"
 
 #include <string>
 #include <vector>
