+++ common/src/Model/RemoveEntityPropertiesQuickFix.h
